mise à niveau demandeController

+++ ProjetBFI/src/main/java/fr/ajcbanque/projetBFI/dto/ClientDTO.java
@@ -1,22 +1,20 @@
 package fr.ajcbanque.projetBFI.dto;
 
-import java.math.BigDecimal;
-
 public class ClientDTO {
-    private Long       id;
-    private String     code;
-    private String     nom;
-    private String     formeJuridique;
-    private String     pays;
-    private String     numCompteBancaire;
-    private BigDecimal ratingInterne;
+    private Long   id;
+    private String code;
+    private String nom;
+    private String formeJuridique;
+    private String pays;
+    private String numCompteBancaire;
+    private Float  ratingInterne;
 
     public ClientDTO() {
 	//
     }
 
     public ClientDTO(Long id, String code, String nom, String formeJuridique,
-	    String pays, String numCompteBancaire, BigDecimal ratingInterne) {
+	    String pays, String numCompteBancaire, Float ratingInterne) {
 	setId(id);
 	setCode(code);
 	setNom(nom);
@@ -74,11 +72,11 @@ public class ClientDTO {
 	this.numCompteBancaire = numCompteBancaire;
     }
 
-    public BigDecimal getRatingInterne() {
+    public Float getRatingInterne() {
 	return ratingInterne;
     }
 
-    public void setRatingInterne(BigDecimal ratingInterne) {
+    public void setRatingInterne(Float ratingInterne) {
 	this.ratingInterne = ratingInterne;
     }
 }
